// libs/petstore-data-access/src/schema.d.ts
export interface paths {
  '/pet': {
    /**
     * Update an existing pet.
     * @description Update an existing pet by Id.
     */
    put: operations['updatePet'];
    /**
     * Add a new pet to the store.
     * @description Add a new pet to the store.
     */
    post: operations['addPet'];
  };
  '/pet/findByStatus': {
    /**
     * Finds Pets by status.
     * @description Multiple status values can be provided with comma separated strings.
     */
    get: operations['findPetsByStatus'];
  };
  '/pet/findByTags': {
    /**
     * Finds Pets by tags.
     * @description Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
     */
    get: operations['findPetsByTags'];
  };
  '/pet/{petId}': {
    /**
     * Find pet by ID.
     * @description Returns a single pet.
     */
    get: operations['getPetById'];
    /**
     * Updates a pet in the store with form data.
     * @description Updates a pet resource based on the form data.
     */
    post: operations['updatePetWithForm'];
    /**
     * Deletes a pet.
     * @description Delete a pet.
     */
    delete: operations['deletePet'];
  };
  '/pet/{petId}/uploadImage': {
    /**
     * Uploads an image.
     * @description Upload image of the pet.
     */
    post: operations['uploadFile'];
  };
  '/store/inventory': {
    /**
     * Returns pet inventories by status.
     * @description Returns a map of status codes to quantities.
     */
    get: operations['getInventory'];
  };
  '/store/order': {
    /**
     * Place an order for a pet.
     * @description Place a new order in the store.
     */
    post: operations['placeOrder'];
  };
  '/store/order/{orderId}': {
    /**
     * Find purchase order by ID.
     * @description For valid response try integer IDs with value <= 5 or > 10. Other values will generate exceptions.
     */
    get: operations['getOrderById'];
    /**
     * Delete purchase order by identifier.
     * @description For valid response try integer IDs with value < 1000. Anything above 1000 or non-integers will generate API errors.
     */
    delete: operations['deleteOrder'];
  };
  '/user': {
    /**
     * Create user.
     * @description This can only be done by the logged in user.
     */
    post: operations['createUser'];
  };
  '/user/createWithList': {
    /**
     * Creates list of users with given input array.
     * @description Creates list of users with given input array.
     */
    post: operations['createUsersWithListInput'];
  };
  '/user/login': {
    /**
     * Logs user into the system.
     * @description Log into the system.
     */
    get: operations['loginUser'];
  };
  '/user/logout': {
    /**
     * Logs out current logged in user session.
     * @description Log user out of the system.
     */
    get: operations['logoutUser'];
  };
  '/user/{username}': {
    /**
     * Get user by user name.
     * @description Get user detail based on username.
     */
    get: operations['getUserByName'];
    /**
     * Update user resource.
     * @description This can only be done by the logged in user.
     */
    put: operations['updateUser'];
    /**
     * Delete user resource.
     * @description This can only be done by the logged in user.
     */
    delete: operations['deleteUser'];
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    Order: {
      /**
       * Format: int64
       * @example 10
       */
      id?: number;
      /**
       * Format: int64
       * @example 198772
       */
      petId?: number;
      /**
       * Format: int32
       * @example 7
       */
      quantity?: number;
      /** Format: date-time */
      shipDate?: string;
      /**
       * @description Order Status
       * @example approved
       * @enum {string}
       */
      status?: 'placed' | 'approved' | 'delivered';
      complete?: boolean;
    };
    Category: {
      /**
       * Format: int64
       * @example 1
       */
      id?: number;
      /** @example Dogs */
      name?: string;
    };
    User: {
      /**
       * Format: int64
       * @example 10
       */
      id?: number;
      /** @example theUser */
      username?: string;
      /** @example John */
      firstName?: string;
      /** @example James */
      lastName?: string;
      /** @example john@email.com */
      email?: string;
      /** @example 12345 */
      password?: string;
      /** @example 12345 */
      phone?: string;
      /**
       * Format: int32
       * @description User Status
       * @example 1
       */
      userStatus?: number;
    };
    Tag: {
      /** Format: int64 */
      id?: number;
      name?: string;
    };
    Pet: {
      /**
       * Format: int64
       * @example 10
       */
      id?: number;
      /** @example doggie */
      name: string;
      category?: components['schemas']['Category'];
      photoUrls: string[];
      tags?: components['schemas']['Tag'][];
      /**
       * @description pet status in the store
       * @enum {string}
       */
      status?: 'available' | 'pending' | 'sold';
    };
    ApiResponse: {
      /** Format: int32 */
      code?: number;
      type?: string;
      message?: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: {
    /** @description Pet object that needs to be added to the store */
    Pet?: {
      content: {
        'application/json': components['schemas']['Pet'];
        'application/xml': components['schemas']['Pet'];
      };
    };
    /** @description List of user object */
    UserArray?: {
      content: {
        'application/json': components['schemas']['User'][];
      };
    };
  };
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {
  /**
   * Update an existing pet.
   * @description Update an existing pet by Id.
   */
  updatePet: {
    /** @description Update an existent pet in the store */
    requestBody: {
      content: {
        'application/json': components['schemas']['Pet'];
        'application/xml': components['schemas']['Pet'];
        'application/x-www-form-urlencoded': components['schemas']['Pet'];
      };
    };
    responses: {
      /** @description Successful operation */
      200: {
        content: {
          'application/json': components['schemas']['Pet'];
          'application/xml': components['schemas']['Pet'];
        };
      };
      /** @description Invalid ID supplied */
      400: {
        content: never;
      };
      /** @description Pet not found */
      404: {
        content: never;
      };
      /** @description Validation exception */
      422: {
        content: never;
      };
      /** @description Unexpected error */
      default: {
        content: never;
      };
    };
  };
// ...
}